recently modified 101-mul

+++ 0x0C-more_malloc_free/101-mul.c
@@ -1,4 +1,5 @@
 #include <stdlib.h>
+#include <stdio.h>
 #include "main.h"
 
 /**
@@ -10,8 +11,8 @@
  */
 int main(int argc, char **argv)
 {
-	int num1, num2;
-	
+	long int num1, num2;
+
 	if (argc != 3)
 	{
 		printf("Error");
@@ -19,7 +20,7 @@ int main(int argc, char **argv)
 	}
 	num1 = atoi(argv[1]);
 	num2 = atoi(argv[2]);
-	printf("%d\n", num1 * num2);
+	printf("%ld\n", num1 * num2);
 
 	return (0);
 }
